Move session configuration to Applications
