+++ portal/src/graphql/portal/ApplicationsConfigurationScreen.module.scss
@@ -5,7 +5,7 @@
 
 .widget {
   margin: 10px 15px;
-  min-width: 600px;
+  max-width: 650px;
 }
 
 .controlGroup {
@@ -21,7 +21,6 @@
 
 .control {
   margin: 10px;
-  width: 600px;
 }
 
 .columnHeader {
