--- portal/src/graphql/portal/SessionConfigurationScreen.module.scss
@@ -1,18 +1,0 @@
-.root {
-  display: flex;
-  flex-direction: column;
-}
-
-.widget {
-  margin: 10px 15px;
-  max-width: 672px;
-}
-
-.controlGroup {
-  display: flex;
-  flex-direction: column;
-}
-
-.control {
-  margin: 10px;
-}
